requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-extensions==3.2.1
django-shell-plus==1.1.7
djangorestframework==3.14.0
executing==1.2.0
filelock==3.11.0
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2023.3.20
platformdirs==3.2.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
